// backend/src/validators/product.schema.ts
import { z } from "zod";

export const createProductSchema = z.object({
  category_id: z.number(),
  subcategory_id: z.number(),
  brand_id: z.number().nullable().optional(),
  name: z.string().min(1),
  slug: z.string().min(1),
  short_description: z.string().optional(),
  description: z.string().optional(),
  sku: z.string().min(1),
  barcode: z.string().optional(),
  price: z.number().positive(),
  discount_price: z.number().positive().optional(),
  tax_percentage: z.number().optional(),
  stock_quantity: z.number().int().min(0).optional(),
  weight: z.number().optional(),
  length: z.number().optional(),
  width: z.number().optional(),
  height: z.number().optional(),
  thumbnail_image: z.string().optional(),
  handle: z.string().optional(),
  title: z.string().optional(),
  option1_name: z.string().optional(),
  option2_name: z.string().optional(),
  option3_name: z.string().optional(),
  hs_code: z.string().optional(),
  country_of_origin: z.string().optional(),
  location: z.string().optional(),
  bin_name: z.string().optional(),
  incoming: z.number().int().optional(),
  unavailable: z.number().int().optional(),
  committed: z.number().int().optional(),
  available: z.number().int().optional(),
  onhand_old: z.number().int().optional(),
  onhand_new: z.number().int().optional(),
  status: z.string().optional(),
  is_featured: z.boolean().optional(),
  is_active: z.boolean().optional(),
});